custom_components/mijnted/requirements_test.txt
pytest
pytest-asyncio
pytest-homeassistant-custom-component
requests
